ukuweb/requirements.txt
django
djangorestframework
psycopg2
django-webpack-loader
python-dateutil
XlsxWriter
django-import-export
requests
